// OneShotJobPortal/src/app/Models/EmployerInfo.ts
export interface EmployerInfo {
    empId: number;
    empCode: string;
    empProfileCreationDate: string;
    empTypeId: number;
    weValueRegDate: string | null;
    weGrowRegDate: string | null;
    incId: number;
    empFullName: string;
    empEmailId: string;
    password: string;
    empContactNo: number;
    empAddress: string;
    empIncorporationDate: string;
    businessStreamId: number;
    empGstin: string;
    empPan: string;
    empDepartment: string;
    empContactPersonName: string;
    empContactPersonNumber: number;
    empContactPersonAltNumber: number | null;
    lastModifiedDate: string;
    companyLogoImage: string;
    fbId: string;
    instaId: string;
    twtrId: string;
    lkdnId: string;
    empWebsite: string;
    empAboutUs: string;
}